Build Bug
Currently dealing with an issue in the build files. When serving the static files with the node back end, the page is completely accessible on the development PC.

On the other hand, when copying the static files over to RaspberryPi the navigation to the livestream page has failed. Everything else in the app, such as the video list page, recording and error modal, is working. However the issue persists with the livestream page.

There have been some instances in which everything works accordingly.
Will try to build the static files in the RPI itself as opposed to using the pre-built ones.